% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{manhattan plots}
\usage{
manhattan.plot(chr, pos, pvalue, sig.level = NA, annotate = NULL,
  ann.default = list(), should.thin = T, thin.pos.places = 2,
  thin.logp.places = 2, xlab = "Chromosome", ylab = expression(-log[10](p
  - value)), col = c("grey39", "darkgray"), panel.extra = NULL, pch = 20,
  cex = 0.8, ...)
}
\description{
manhattan plots
}
