% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{procdist.array}
\alias{procdist.array}
\title{Calculates procrustes distance for array format}
\usage{
procdist.array(data, reference)
}
\arguments{
\item{data}{A set of data you want to get distances for}

\item{reference}{some reference you want every observation to be measured against}
}
\value{
Procrustes distances
}
\description{
Calculates procrustes distance for array format
}
