% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{procdist}
\alias{procdist}
\title{Calculates procrustes distance for row format}
\usage{
procdist(data, target)
}
\arguments{
\item{data}{A set of data you want to get distances for}

\item{target}{some reference you want every observation to be measured against}
}
\value{
Procrustes distances
}
\description{
Calculates procrustes distance for row format
}
