% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{row2array}
\alias{row2array}
\title{Converts row formats to array format for 2D data}
\usage{
row2array(data, Nlandmarks = length(data[1, ])/2)
}
\value{
An array
}
\description{
Converts row formats to array format for 2D data
}
