% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{row2array3d}
\alias{row2array3d}
\title{Converts row formats to array format for 3D data}
\usage{
row2array3d(data, Nlandmarks = length(data[1, ])/3)
}
\value{
An array
}
\description{
Converts row formats to array format for 3D data
}
